- move console logic to console tab - colorize logs when color is set
